requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.1
